scripts/scraper/requirements.txt
beautifulsoup4==4.12.3
selenium==4.27.1
requests==2.32.3
psycopg2-binary==2.9.10
pandas==2.2.3
pdfplumber==0.11.4
schedule==1.2.2
python-dotenv==1.0.1
lxml==5.3.0
undetected-chromedriver==3.5.5
geopandas==1.0.1
